Rename best_tip_height module to chain_tip
